LayoutTests/fast/dom/HTMLProgressElement/progress-clone-expected.txt

PASS cloned.value is target.value
PASS layoutTestController.shadowPseudoId(clonedShadowRoot.firstChild) is layoutTestController.shadowPseudoId(targetShadowRoot.firstChild)
PASS layoutTestController.shadowPseudoId(clonedShadowRoot.firstChild.firstChild) is layoutTestController.shadowPseudoId(targetShadowRoot.firstChild.firstChild)
PASS clonedShadowRoot.firstChild.firstChild.style.width is "70%"
PASS targetShadowRoot.firstChild.firstChild.style.width is "50%"
PASS successfullyParsed is true

TEST COMPLETE
